requirements.txt
fastapi
uvicorn[standard]
python-multipart
httpx
twilio
elevenlabs
anthropic
langgraph
langchain
langchain-anthropic
langchain-community
faiss-cpu
supabase
celery
redis
upstash-redis
scikit-learn
xgboost
joblib
google-cloud-vision
gspread
google-auth
pandas
numpy
python-dotenv
openai
pytesseract
Pillow
groq
cryptography
